SM-58 moved mock to assets, added translation to modal
